+++ simple_history/middleware.py
@@ -13,17 +13,9 @@ class HistoryRequestMiddleware:
         self.get_response = get_response
 
     def __call__(self, request):
-        # Code to be executed for each request before
-        # the view (and later middleware) are called.
-
         HistoricalRecords.thread.request = request
-
-        response = self.get_response(request)
-
-        # Code to be executed for each request/response after
-        # the view is called.
-
-        if hasattr(HistoricalRecords.thread, 'request'):
-            del HistoricalRecords.thread.request
-
-        return response
+        try:
+            return self.get_response(request)
+        finally:
+            if hasattr(HistoricalRecords.thread, 'request'):
+                del HistoricalRecords.thread.request