+++ simple_history/tests/tests/test_middleware.py
@@ -4,6 +4,7 @@ from django.conf import settings
 from django.test import TestCase, override_settings
 from django.urls import reverse
 
+from simple_history.models import HistoricalRecords
 from simple_history.tests.custom_user.models import CustomUser
 from simple_history.tests.models import Poll
 
@@ -13,6 +14,7 @@ overridden_settings = {
 }
 
 
+@override_settings(**overridden_settings)
 class MiddlewareTest(TestCase):
     def setUp(self):
         self.user = CustomUser.objects.create_superuser(
@@ -21,7 +23,6 @@ class MiddlewareTest(TestCase):
             'pass'
         )
 
-    @override_settings(**overridden_settings)
     def test_user_is_set_on_create_view_when_logged_in(self):
         self.client.force_login(self.user)
         data = {
@@ -37,7 +38,6 @@ class MiddlewareTest(TestCase):
         self.assertListEqual([ph.history_user_id for ph in poll_history],
                              [self.user.id])
 
-    @override_settings(**overridden_settings)
     def test_user_is_not_set_on_create_view_not_logged_in(self):
         data = {
             'question': 'Test question',
@@ -52,7 +52,6 @@ class MiddlewareTest(TestCase):
         self.assertListEqual([ph.history_user_id for ph in poll_history],
                              [None])
 
-    @override_settings(**overridden_settings)
     def test_user_is_set_on_update_view_when_logged_in(self):
         self.client.force_login(self.user)
         poll = Poll.objects.create(
@@ -76,7 +75,6 @@ class MiddlewareTest(TestCase):
         self.assertListEqual([ph.history_user_id for ph in poll_history],
                              [self.user.id, None])
 
-    @override_settings(**overridden_settings)
     def test_user_is_not_set_on_update_view_when_not_logged_in(self):
         poll = Poll.objects.create(
             question='Test question',
@@ -99,7 +97,6 @@ class MiddlewareTest(TestCase):
         self.assertListEqual([ph.history_user_id for ph in poll_history],
                              [None, None])
 
-    @override_settings(**overridden_settings)
     def test_user_is_unset_on_update_view_after_logging_out(self):
         self.client.force_login(self.user)
         poll = Poll.objects.create(
@@ -137,7 +134,6 @@ class MiddlewareTest(TestCase):
         self.assertListEqual([ph.history_user_id for ph in poll_history],
                              [None, self.user.id, None])
 
-    @override_settings(**overridden_settings)
     def test_user_is_set_on_delete_view_when_logged_in(self):
         self.client.force_login(self.user)
         poll = Poll.objects.create(
@@ -155,7 +151,6 @@ class MiddlewareTest(TestCase):
         self.assertListEqual([ph.history_user_id for ph in poll_history],
                              [self.user.id, None])
 
-    @override_settings(**overridden_settings)
     def test_user_is_not_set_on_delete_view_when_not_logged_in(self):
         poll = Poll.objects.create(
             question='Test question',
@@ -171,3 +166,8 @@ class MiddlewareTest(TestCase):
 
         self.assertListEqual([ph.history_user_id for ph in poll_history],
                              [None, None])
+
+    def test_exception_handling(self):
+        with self.assertRaises(ValueError):
+            self.client.get(reverse('raise-exception'))
+        self.assertFalse(hasattr(HistoricalRecords.thread, 'request'))