+++ simple_history/tests/tests/test_models.py
@@ -53,7 +53,10 @@ from ..models import (
     UnicodeVerboseName,
     UUIDModel,
     UUIDDefaultModel,
-    WaterLevel
+    WaterLevel,
+    TextFieldChangeReasonModel1,
+    TextFieldChangeReasonModel2,
+    CharFieldChangeReasonModel,
 )
 
 get_model = apps.get_model
@@ -381,6 +384,57 @@ class HistoricalRecordsTest(TestCase):
 
         history = entry.history.all()[0]
         self.assertTrue(isinstance(history.history_id, uuid.UUID))
+
+    def test_default_history_change_reason(self):
+        entry = CharFieldChangeReasonModel.objects.create(
+            greeting="what's up?"
+        )
+        history = entry.history.get()
+
+        self.assertEqual(history.history_change_reason, None)
+
+    def test_charfield_history_change_reason(self):
+        # Default CharField and length
+        entry = CharFieldChangeReasonModel.objects.create(
+            greeting="what's up?"
+        )
+        entry.greeting = "what is happening?"
+        entry.save()
+        update_change_reason(entry, 'Change greeting.')
+
+        history = entry.history.all()[0]
+        field = history._meta.get_field('history_change_reason')
+
+        self.assertTrue(isinstance(field, models.CharField))
+        self.assertTrue(field.max_length, 100)
+
+    def test_textfield_history_change_reason1(self):
+        # TextField usage is determined by settings
+        entry = TextFieldChangeReasonModel1.objects.create(
+            greeting="what's up?"
+        )
+        entry.greeting = "what is happening?"
+        entry.save()
+        update_change_reason(entry, 'Change greeting.')
+
+        history = entry.history.all()[0]
+        field = history._meta.get_field('history_change_reason')
+
+        self.assertTrue(isinstance(field, models.TextField))
+
+    def test_textfield_history_change_reason2(self):
+        # TextField instance is passed in init
+        entry = TextFieldChangeReasonModel2.objects.create(
+            greeting="what's up?"
+        )
+        entry.greeting = "what is happening?"
+        entry.save()
+        update_change_reason(entry, 'Change greeting.')
+
+        history = entry.history.all()[0]
+        field = history._meta.get_field('history_change_reason')
+
+        self.assertTrue(isinstance(field, models.TextField))
 
     def test_get_prev_record(self):
         poll = Poll(question="what's up?", pub_date=today)