+++ CHANGES.rst
@@ -1,8 +1,8 @@
 Changes
 =======
 
-Unreleased
-----------
+2.1.0 (2018-06-04)
+------------------
 - Add ability to specify custom history_reason field (gh-379)
 - Add ability to specify custom history_id field (gh-368)
 - Add HistoricalRecord instance properties `prev_record` and `next_record` (gh-365)